Allow additionalProperties to take Boolean values
The OpenAPI specification defines the additionalProperties field
according to the JSON schema specification. In the version of the JSON
schema specification referenced by the OpenAPI specification [1], the
value of additionalProperties may be either Boolean or a schema (though
it appears that newer versions of the JSON schema specification [2] no
longer allow Boolean values).

However, to comply with OpenAPI 3.0, this adjusts the representation of
the additionalProperties field in this crate accordingly

[1] https://tools.ietf.org/html/draft-wright-json-schema-validation-00#section-5.18
[2] https://json-schema.org/draft/2019-09/json-schema-core.html#additionalProperties