Replace failure dependency with thiserror/anyhow
The failure crate appears to be deprecated these days and has drawbacks
including incompatibilities with std::error::Error. For users, it might
be inconvenient to have to deal with error types that don’t implement
std::error::Error.

For this reason, this replaces the failure crate entirely. In the
library part, thiserror is used instead, which also makes the generation
of the error type short and convenient (even taking care of the From
implementations), without exposing any thiserror specifics to the user.

In the example, anyhow now takes care of printing nicely formatted error
messages, including error causes and (if provided with RUST_BACKTRACE=1)
an error backtrace.